+++ SurvivalGame/Assets/WorldObjects/Chest.cs
@@ -12,7 +12,7 @@ public class Chest : Building<BuildingData>, IUsableBuilding
         base.Start();
     }
 
-    override public bool InteractWithBuilding()
+    override public bool InteractWithWorldObject()
     {
         return Use();
     }