+++ CHANGELOG.md
@@ -163,3 +163,7 @@
 - Added rudimental shader to fade objects that are between camera and player (Fading/See-through shader tecnique with dithering).
 - Rudimental logic for `Resource` objects to produce items at end of day (e.g. grow bananas on trees).
 - Some new resources and item UI icons.
+---
+16-3-2025
+- Added logic to harvest produced items from `Resource` world objects.
+- Refactored `Resource` class producing and spawning items logics.